pyproject.toml
[project]
name = "splatter-ai-powered-object-scatter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cython>=3.1.3",
    "fake-bpy-module-latest>=20250630",
    "ninja>=1.13.0",
    "numpy>=2.2.6",
]
